testing/web-platform/mozilla/tests/indic-detection/tab.html

<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
™кхҐх£тҐ (Mars) ≈у§тиҐ∞ісҐроҐо§фҐ ачҐч еЉ Ђи£чҐ Ё∞сҐ. ёґ ≈у§тщ§ф§ЉпҐґ п£щҐи£хґ Ђи£ч£и ачҐчґ.ёиҐ∞ісҐроҐо§фҐ с§икҐк§ш§т Ђи£ч£щ єо¬иҐ∞ №іоҐоо£и ёунҐм£хґ к§ш§т Ђи£ч£и ™кхҐх£тҐ ёЉиҐи§шґ. Ђсщ£мҐЃщуҐ ёиҐЂи£јиҐ∞ Ђр£уҐиҐимЊч§щҐ ™ртђукҐ ≈мҐЃїчҐчщуҐ. ёощҐ ЂсшҐрурҐр§фҐ и£нрҐрісҐ ёЉсҐє ЁиҐђкі ёиҐЂи£ђчкҐ ™кпҐп§шс£ииҐ и£мҐіи§шґ.[12] ёощ£ЂфЂт ёошҐ∞кҐ ™кхҐх£тҐ вщҐш ™ртуҐ гшҐрмҐмґ. еЉ єх§к£уҐ Ђи£ч£щ ёощҐ ЂсшҐрурҐє кпҐо§ущ§фҐ ачҐчґЂр£фҐ и§нҐниҐ ∞ц§иђчїсҐ, єх§т§фҐ ачҐчґ Ђр£щҐш ву§сђфичҐ, рчҐчоҐо£иҐ∞ичҐ, р£ђфхщйҐичҐ, рщ§ЌЃт ґЉхрҐ р∞о§иђчїсҐ ™и£нҐмґ. ™кхҐх£т§щҐ ≤цшҐк§иҐи£фЇсҐ, рЉх с£шҐшйҐијсҐ єх§иҐ∞ ачҐчђорҐ Ђр£щҐшђхЂт. ≈у§т снҐмфоҐґчҐ с§и атус£щ еф§сҐр≤ сђфїсҐ, с§ирҐ™ру§т ™кйҐ∞оҐґрҐ рчҐчоҐо£иҐ∞ијчҐ ещҐш£щ су§щуҐ рчҐчоҐо£иҐ∞сҐ ™кхҐх£т§ЂфЂт ачҐчщ.
<script>
test(function() {
  assert_equals(document.characterSet, "windows-1252");
},"Should fall back to windows-1252");
</script>
The text content above is a converted extract from the start of the Wikipedia article <a href="https://ta.wikipedia.org/w/index.php?title=%E0%AE%9A%E0%AF%86%E0%AE%B5%E0%AF%8D%E0%AE%B5%E0%AE%BE%E0%AE%AF%E0%AF%8D_(%E0%AE%95%E0%AF%8B%E0%AE%B3%E0%AF%8D)&oldid=3129711">&#2970;&#3014;&#2997;&#3021;&#2997;&#3006;&#2991;&#3021; (&#2965;&#3019;&#2995;&#3021;)</a> and
is licensed under the <a href="https://creativecommons.org/licenses/by-sa/3.0/legalcode">Creative Commons Attribution-ShareAlike 3.0 Unported</a> license.
